refactor(rubsun): add storage wrapper module to remove direct Room dependency

+++ features/rubsun/api/src/main/java/com/rubsun/api/domain/models/NumberDisplayModel.kt
@@ -3,7 +3,7 @@ package com.rubsun.api.domain.models
 data class NumberDisplayModel(
     val value: Int,
     val label: String,
-    val squared: Int,
+    val squared: Long,
 )
